requirements.txt
openai
jinja2
weasyprint
python-dotenv
flask
gspread
oauth2client